+++ examples/react_demo/src/JSONEditorDemo.js
@@ -9,8 +9,7 @@ export default class JSONEditorDemo extends Component {
   componentDidMount () {
     const options = {
       mode: 'tree',
-      modes: ['tree', 'code'],
-      onChange: this.onChange
+      onChangeJSON: this.props.onChangeJSON
     };
 
     this.jsoneditor = new JSONEditor(this.container, options);
@@ -31,14 +30,5 @@ export default class JSONEditorDemo extends Component {
     return (
         <div className="jsoneditor-react-container" ref={elem => this.container = elem} />
     );
-  }
-
-  onChange = () => {
-    if (this.props.onChange) {
-      // note that this.jsoneditor.get() can fail in mode text/code
-      // when the contents is no valid JSON object.
-      // So if you need mode text/node, you must use getText() and setText().
-      this.props.onChange(this.jsoneditor.get());
-    }
   }
 }